Polling as an action?!
